# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Testing!", "version": "0.1.0"}, "paths": {"/api/user": {"post": {"tags": ["user"], "summary": "Create User", "operationId": "create_user", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/user/{user_id}": {"get": {"tags": ["user"], "summary": "Get User By Id", "operationId": "get_user_by_id", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["user"], "summary": "Delete User By Id", "operationId": "delete_user_by_id", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["user"], "summary": "Update User By Id", "operationId": "update_user_by_id", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users": {"get": {"tags": ["user"], "summary": "Get User By Name", "operationId": "get_user_by_name", "parameters": [{"name": "user_name", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "User Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["user"], "summary": "Delete User By Name", "operationId": "delete_user_by_name", "parameters": [{"name": "user_name", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "User Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/me": {"get": {"tags": ["user"], "summary": "Get Current User", "operationId": "get_current_user", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/User"}, {"type": "null"}], "title": "Response Get Current User Api Users Me Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/get-token": {"post": {"tags": ["user"], "summary": "Generate Token", "operationId": "generate_token", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_generate_token_api_get_token_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/add-favorite": {"get": {"tags": ["user"], "summary": "Add Favorite Book", "operationId": "add_favorite_book", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "book_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Book Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FavoriteBookLink"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/remove-favorite": {"get": {"tags": ["user"], "summary": "Remove Favorite Book", "operationId": "remove_favorite_book", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "book_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Book Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FavoriteBookLink"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/get-favorites": {"get": {"tags": ["user"], "summary": "Get Favorite Books", "operationId": "get_favorite_books", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/BookView"}, "type": "array", "title": "Response Get Favorite Books Api Users Get Favorites Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/upload-book/": {"post": {"tags": ["book"], "summary": "Upload Book", "operationId": "upload_book", "parameters": [{"name": "name", "in": "query", "required": true, "schema": {"type": "string", "title": "Name"}}, {"name": "description", "in": "query", "required": true, "schema": {"type": "string", "title": "Description"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_book_api_upload_book__post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Book"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/get-book/{id}": {"get": {"tags": ["book"], "summary": "Get Book By Id", "operationId": "get_book_by_id", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/Book"}, {"type": "null"}], "title": "Response Get Book By Id Api Get Book  Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/get-book-full/{id}": {"get": {"tags": ["book"], "summary": "Get Book By Id With Favorite And Reviews", "operationId": "get_book_by_id_with_favorite_and_reviews", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookViewReview"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/get-books": {"get": {"tags": ["book"], "summary": "Get Books", "operationId": "get_books", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "offset", "in": "query", "required": true, "schema": {"type": "integer", "title": "Offset"}}, {"name": "limit", "in": "query", "required": true, "schema": {"type": "integer", "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/BookView"}, "title": "Response Get Books Api Get Books Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/files/{file_path}": {"get": {"tags": ["book"], "summary": "Read File", "operationId": "read_file", "parameters": [{"name": "file_path", "in": "path", "required": true, "schema": {"type": "string", "title": "File Path"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/review": {"post": {"tags": ["review"], "summary": "Create Review", "operationId": "create_review", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Review"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/get-reviews-book-id": {"get": {"tags": ["review"], "summary": "Get Reviews By Book Id", "operationId": "get_reviews_by_book_id", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "book_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "Book Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Review"}, "title": "Response Get Reviews By Book Id Api Get Reviews Book Id Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api": {"get": {"summary": "Root", "operationId": "root", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Root Api Get"}}}}}}}}, "components": {"schemas": {"Body_generate_token_api_get_token_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_generate_token_api_get_token_post"}, "Body_upload_book_api_upload_book__post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File", "description": "A file read as UploadFile"}}, "type": "object", "required": ["file"], "title": "Body_upload_book_api_upload_book__post"}, "Book": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "file_path": {"type": "string", "title": "File Path"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}}, "type": "object", "required": ["name", "description", "file_path", "id"], "title": "Book"}, "BookView": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "file_path": {"type": "string", "title": "File Path"}, "id": {"type": "integer", "title": "Id"}, "isFavorite": {"type": "boolean", "title": "Isfavorite"}}, "type": "object", "required": ["name", "description", "file_path", "id", "isFavorite"], "title": "BookView"}, "BookViewReview": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "file_path": {"type": "string", "title": "File Path"}, "id": {"type": "integer", "title": "Id"}, "isFavorite": {"type": "boolean", "title": "Isfavorite"}, "reviews": {"items": {"$ref": "#/components/schemas/Review"}, "type": "array", "title": "Reviews"}}, "type": "object", "required": ["name", "description", "file_path", "id", "isFavorite", "reviews"], "title": "BookViewReview"}, "FavoriteBookLink": {"properties": {"user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "User Id"}, "book_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Book Id"}}, "type": "object", "title": "FavoriteBookLink"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Review": {"properties": {"content": {"type": "string", "title": "Content"}, "owner_id": {"type": "integer", "title": "Owner Id"}, "book_id": {"type": "integer", "title": "Book Id"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}}, "type": "object", "required": ["content", "id"], "title": "Review"}, "ReviewCreate": {"properties": {"content": {"type": "string", "title": "Content"}, "owner_id": {"type": "integer", "title": "Owner Id"}, "book_id": {"type": "integer", "title": "Book Id"}}, "type": "object", "required": ["content"], "title": "ReviewCreate"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "User": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "pass_hash": {"type": "string", "title": "Pass Hash"}}, "type": "object", "required": ["username", "email", "id", "pass_hash"], "title": "User"}, "UserCreate": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "email", "password"], "title": "UserCreate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/get-token"}}}}}}